refactor: #8 各種コンポーネントのディレクトリ配置を変更。
