+++ bear-clone/src/app/shared/components/editor/editor.component.spec.ts
