+++ bear-clone/src/app/shared/components/note-list/note-item/note-item.component.spec.ts
