+++ bear-clone/src/app/shared/components/note-list/note-list.component.html
