+++ bear-clone/src/app/shared/components/note-list/search-box/search-box.component.html
