+++ bear-clone/src/app/shared/components/sidebar/sidebar.component.spec.ts
