+++ bear-clone/src/app/trash/pages/trash-list/trash-list.component.html
