+++ bear-clone/src/app/trash/pages/trash-list/trash-list.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from "@angular/core";
-import { NoteListComponent } from "../note-list/note-list.component";
+import { NoteListComponent } from "../../../shared/components/note-list/note-list.component";
 import { Store } from "@ngrx/store";
-import { selectAllDeletedNotes } from "../shared/selectors/note.selectors";
+import { selectAllDeletedNotes } from "../../../shared/selectors/note.selectors";
 
 @Component({
   selector: "app-trash-list",
